The niceification continues.
I can't remember what all I've done to this now, but everything is just nicer and easier to use. I've also reviewed the code a fair amount while niceifying it, and feel more capable of improving the disassembler. Though, I may put a pin in doing that completionist decode just so I can proceed to the next part of the course.