Moving along
Since most of the assembler code is a year old at this point, I've been playing around with doing a full rewrite (in the new remake_ folder). But it would probably be smarter to just move on to the next part of the course. Then, if I need to come back and make the decoder GOOD, we'll cross that bridge when we get to it.