Simulation is Listing 0044 compliant
I gave up on the disasm rewrite, which was the right call because we're moving on to the sim stuff which is already a lot of fun. Right now, sim8086 can execute the assembled Listing 44 without issues. The code's a little hacky, and logging is probably TOO verbose. But I'll make it better later. For now, it's 3:23 AM...