Decode rewrite
I fully rewrote the decode logic into a header-only library, and the new sim8086.c program uses it to both simulate and disassemble. My coverage of .asm operations has not improved since the last commit, but we should be in a much better place to add simulation and disassembling features simultaneously from now on. Would love to go back and get it all working with that completionist decode some day! I also added FLAGS and IP registers, but they don't do anything yet.