Rectangles are cancelled!
Casey's video on drawing rectangles is more of a cool example than a learning exercise. He said implementing it can be skipped, and I unterstood exactly how everything in that video worked, so I'm not going to bother implementing file writing. Instead I've nicified the code a little, and am ready to move past the 8086 exercises!