On to part 2
I haven't done all of the 8086 exercises, but I understand the material now and want to move on to part 2 of the course. So far, I've made a basic .json generator, and the first order of business is to improve it so I can profile it in the coming exercises.